refactor: Phase 2 — extract query-prolog.ts and prolog-facts.ts
- Move Prolog query builders (resolveParentPredicate, buildParentQuery,
  buildWhereQuery, buildOrderQuery, etc.) into query-prolog.ts
- Move generatePrologFacts() + helpers (toSnakeCase, buildInstanceClause,
  buildPropertyClause, buildRelationClause) into prolog-facts.ts
- Remove duplicate capitalize() from Ad4mModel.ts (already in util.ts)
- Ad4mModel.ts re-exports generatePrologFacts for backward compat
- Update barrel index.ts with new modules

All 119 tests pass (115 passed, 4 skipped).
