+++ core/src/model/Ad4mModel.ts
@@ -3,9 +3,10 @@ import { Link } from "../links/Links";
 import { LinkQuery } from "../perspectives/LinkQuery";
 import { PerspectiveProxy } from "../perspectives/PerspectiveProxy";
 import { makeRandomId, PropertyOptions, Model, getPropertiesMetadata, getRelationsMetadata, setPropertyRegistryEntry, setRelationRegistryEntry, PropertyMetadataEntry, RelationMetadataEntry, buildConformanceFilter } from "./decorators";
-import { singularToPlural, pluralToSingular, propertyNameToSetterName, relationToAdderName, relationToRemoverName, relationToSetterName } from "./util";
+import { singularToPlural, pluralToSingular, capitalize, propertyNameToSetterName, relationToAdderName, relationToRemoverName, relationToSetterName } from "./util";
 import { escapeSurrealString } from "../utils";
 import { formatSurrealValue as _formatSurrealValue, compileWhereClause } from "./surreal-utils";
+import { resolveParentPredicate, buildParentQuery, buildAuthorAndTimestampQuery, buildPropertiesQuery, buildWhereQuery, buildCountQuery, buildOrderQuery, buildOffsetQuery, buildLimitQuery } from "./query-prolog";
 import type {
   WhereCondition, Where, Order, ParentScope, IncludeMap, Query,
   RelationSubQuery, GetOptions, AllInstancesResult, ResultsWithTotalCount,
@@ -52,190 +53,6 @@ interface JSONSchemaToModelOptions {
   resolveLanguage?: string;
   local?: boolean;
   propertyOptions?: Record<string, Partial<PropertyOptions>>;
-}
-
-function capitalize(word: string): string {
-  return word.charAt(0).toUpperCase() + word.slice(1);
-}
-
-/**
- * Resolves the predicate for a parent query.
- *
- * Uses TS discriminated union narrowing:
- * - Raw form (`{ id, predicate }`) → predicate used as-is
- * - Model form (`{ model, id, field? }`) → lookup from relation metadata
- *   - With `field`: direct key lookup
- *   - Without `field`: scan for a relation whose `target()` matches `childCtor`
- */
-function resolveParentPredicate(
-  parent: ParentScope,
-  childCtor: Function,
-): string {
-  // Raw form — explicit predicate
-  if ('predicate' in parent) return parent.predicate;
-
-  // Model form — resolve from relation metadata
-  const { model } = parent;
-  const relMeta = getRelationsMetadata(model);
-
-  // Direct lookup by field name when provided
-  if (parent.field) {
-    const entry = relMeta[parent.field];
-    if (!entry) {
-      throw new Error(
-        `parent(): field "${parent.field}" is not a registered relation on ${model.name}`,
-      );
-    }
-    return entry.predicate;
-  }
-
-  // Fallback: scan for a relation whose target matches the child class
-  for (const [, entry] of Object.entries(relMeta)) {
-    if (entry.target && entry.target() === childCtor) {
-      return entry.predicate;
-    }
-  }
-  throw new Error(
-    `parent(): could not resolve predicate — no relation on ${model.name} targets ${(childCtor as any).name || 'the queried class'}`,
-  );
-}
-
-function buildParentQuery(
-  parent: ParentScope | undefined,
-  resolvedPredicate?: string,
-): string {
-  if (!parent || !resolvedPredicate) return '';
-  return `triple("${parent.id}", "${resolvedPredicate}", Base)`;
-}
-
-// todo: only return Timestamp & Author from query (Base, AllLinks, and SortLinks not required)
-function buildAuthorAndTimestampQuery(): string {
-  // Gets the author and timestamp of a Ad4mModel instance (based on the first link mentioning the base)
-  return `
-    findall(
-      [T, A],
-      link(Base, _, _, T, A),
-      AllLinks
-    ),
-    sort(AllLinks, SortedLinks),
-    SortedLinks = [[Timestamp, Author]|_]
-  `;
-}
-
-function buildPropertiesQuery(properties?: string[]): string {
-  // Gets the name, value, and resolve boolean for all (or some) properties on a Ad4mModel instance
-  // Resolves literals (if property_resolve/2 is true) to their value - either the data field if it is
-  // an Expression in JSON literal, or the direct literal value if it is a simple literal
-  // If no properties are provided, all are included
-  return `
-    findall([PropertyName, PropertyValue, Resolve], (
-      % Constrain to specified properties if provided
-      ${properties ? `member(PropertyName, [${properties.map((name) => `"${name}"`).join(", ")}]),` : ""}
-      resolve_property(SubjectClass, Base, PropertyName, PropertyValue, Resolve)
-    ), Properties)
-  `;
-}
-
-function buildWhereQuery(where: Where = {}): string {
-  // Constrains the query to instances that match the provided where conditions
-  // 'id' maps to the Prolog 'Base' variable (the base expression of the instance).
-
-  function formatValue(value) {
-    // Wrap strings in quotes
-    return typeof value === "string" ? `"${value}"` : value;
-  }
-
-  return (Object.entries(where) as [string, WhereCondition][])
-    .map(([key, value]) => {
-      const isSpecial = ["id", "author", "timestamp"].includes(key);
-      const getter = `resolve_property(SubjectClass, Base, "${key}", Value${key}, _)`;
-      // For 'id' the Prolog variable is always 'Base'
-      const field = key === "id" ? "Base" : capitalize(key);
-
-      // Handle direct array values (for IN conditions)
-      if (Array.isArray(value)) {
-        const formattedValues = value.map((v) => formatValue(v)).join(", ");
-        if (isSpecial) return `member(${field}, [${formattedValues}])`;
-        else return `${getter}, member(Value${key}, [${formattedValues}])`;
-      }
-
-      // Handle operation object
-      if (typeof value === "object" && value !== null) {
-        const { not, between, lt, lte, gt, gte } = value;
-
-        // Handle NOT operation
-        if (not !== undefined) {
-          if (Array.isArray(not)) {
-            // NOT IN array
-            const formattedValues = not.map((v) => formatValue(v)).join(", ");
-            if (isSpecial) return `\\+ member(${field}, [${formattedValues}])`;
-            else return `${getter}, \\+ member(Value${key}, [${formattedValues}])`;
-          } else {
-            // NOT EQUAL
-            if (isSpecial) return `${field} \\= ${formatValue(not)}`;
-            else return `${getter}, Value${key} \\= ${formatValue(not)}`;
-          }
-        }
-
-        // Handle BETWEEN
-        if (between !== undefined && Array.isArray(between) && between.length === 2) {
-          if (isSpecial) return `${field} >= ${between[0]}, ${field} =< ${between[1]}`;
-          else return `${getter}, Value${key} >= ${between[0]}, Value${key} =< ${between[1]}`;
-        }
-
-        // Handle lt, lte, gt, & gte operations
-        const operators = [
-          { value: lt, symbol: "<" }, // LESS THAN
-          { value: lte, symbol: "=<" }, // LESS THAN OR EQUAL TO
-          { value: gt, symbol: ">" }, // GREATER THAN
-          { value: gte, symbol: ">=" }, // GREATER THAN OR EQUAL TO
-        ];
-
-        for (const { value, symbol } of operators) {
-          if (value !== undefined)
-            return isSpecial ? `${field} ${symbol} ${value}` : `${getter}, Value${key} ${symbol} ${value}`;
-        }
-      }
-
-      // Default to direct equality
-      if (isSpecial) return `${field} = ${formatValue(value)}`;
-      else return `${getter}, Value${key} = ${formatValue(value)}`;
-    })
-    .join(", ");
-}
-
-function buildCountQuery(count?: boolean): string {
-  return count ? "length(UnsortedInstances, TotalCount)" : "";
-}
-
-function buildOrderQuery(order?: Order): string {
-  if (!order) return "SortedInstances = UnsortedInstances";
-  const entries = Object.entries(order);
-  if (entries.length === 1) {
-    const [propertyName, direction] = entries[0];
-    return `sort_instances(UnsortedInstances, "${propertyName}", "${direction}", SortedInstances)`;
-  }
-  // Multi-field sort: sort from least-significant to most-significant key
-  // so that the final (primary) sort preserves secondary-key ordering for equal values.
-  // The merge_sort implementation is stable (equal elements keep original order).
-  const clauses: string[] = [];
-  for (let i = entries.length - 1; i >= 0; i--) {
-    const [propertyName, direction] = entries[i];
-    const inputVar = i === entries.length - 1 ? "UnsortedInstances" : `MultiSortIntermediate${i + 1}`;
-    const outputVar = i === 0 ? "SortedInstances" : `MultiSortIntermediate${i}`;
-    clauses.push(`sort_instances(${inputVar}, "${propertyName}", "${direction}", ${outputVar})`);
-  }
-  return clauses.join(",\n      ");
-}
-
-function buildOffsetQuery(offset?: number): string {
-  if (!offset || offset < 0) return "InstancesWithOffset = SortedInstances";
-  return `skipN(SortedInstances, ${offset}, InstancesWithOffset)`;
-}
-
-function buildLimitQuery(limit?: number): string {
-  if (!limit || limit < 0) return "AllInstances = InstancesWithOffset";
-  return `takeN(InstancesWithOffset, ${limit}, AllInstances)`;
 }
 
 function normalizeNamespaceString(namespace: string): string {
@@ -4435,135 +4252,5 @@ export class ModelQueryBuilder<T extends Ad4mModel> {
   }
 }
 
-// ── Standalone Prolog fact generator ──────────────────────────────────────────
-
-/**
- * Convert a camelCase or PascalCase identifier to snake_case.
- * Examples: "TestPost" -> "test_post", "createdAt" -> "created_at"
- */
-function toSnakeCase(str: string): string {
-  return str
-    .replace(/([A-Z])/g, "_$1")
-    .toLowerCase()
-    .replace(/^_/, "");
-}
-
-function buildInstanceClause(predicateName: string, metadata: ModelMetadata): string | null {
-  const props = metadata.properties;
-  // Collect flags first — these are the strongest recognizers
-  const flags = Object.values(props).filter((p) => p.flag && p.predicate && p.initial);
-  if (flags.length > 0) {
-    const conditions = flags
-      .map((p) => `triple(X, '${p.predicate}', '${p.initial}')`)
-      .join(",\n    ");
-    return `${predicateName}(X) :-\n    ${conditions}.`;
-  }
-  // Fallback: required non-flag properties
-  const required = Object.values(props).filter((p) => p.required && p.predicate && !p.flag);
-  if (required.length > 0) {
-    const conditions = required
-      .map((p) => `triple(X, '${p.predicate}', _)`)
-      .join(",\n    ");
-    return `${predicateName}(X) :-\n    ${conditions}.`;
-  }
-  // Open-world structural fallback: match nodes that have at least one of the
-  // model's declared property predicates.  Uses disjunction (;) so any single
-  // predicate is enough.
-  const allPredicates = Object.values(props)
-    .filter((p) => p.predicate)
-    .map((p) => p.predicate!);
-  if (metadata.relations) {
-    for (const rel of Object.values(metadata.relations)) {
-      if (rel.predicate) allPredicates.push(rel.predicate);
-    }
-  }
-  if (allPredicates.length > 0) {
-    const disjunction = allPredicates
-      .map((pred) => `triple(X, '${pred}', _)`)
-      .join(" ;\n    ");
-    return `${predicateName}(X) :-\n    (${disjunction}).`;
-  }
-  return null;
-}
-
-function buildPropertyClause(modelPredicateName: string, prop: PropertyMetadata): string | null {
-  if (prop.flag) return null;
-  if (!prop.predicate) return null;
-  const clauseName = `${modelPredicateName}_${toSnakeCase(prop.name)}`;
-  return `${clauseName}(X, Value) :- triple(X, '${prop.predicate}', Value).`;
-}
-
-function buildRelationClause(modelPredicateName: string, relation: RelationMetadata): string | null {
-  if (!relation.predicate) return null;
-  const clauseName = `${modelPredicateName}_${toSnakeCase(relation.name)}`;
-  if (relation.direction === "reverse") {
-    return `${clauseName}(X, Values) :- findall(V, triple(V, '${relation.predicate}', X), Values).`;
-  }
-  return `${clauseName}(X, Values) :- findall(V, triple(X, '${relation.predicate}', V), Values).`;
-}
-
-/**
- * Generate Prolog predicate facts from a model class's decorator metadata.
- *
- * Given a model class decorated with `@Model` (and its `@Flag`, `@Property`,
- * `@HasMany`, `@BelongsToMany` decorators), this function emits a string of
- * Prolog clauses that can be prepended to any `perspective.infer()` call.
- *
- * The generated predicates are:
- * - **Instance recognizer** — `modelName(X)` — matches instances of the model
- * - **Property getters** — `modelName_propName(X, Value)` — one per property
- * - **Relation getters** — `modelName_relName(X, Values)` — one per relation
- *
- * @example
- * ```typescript
- * import { generatePrologFacts } from '@coasys/ad4m';
- *
- * const facts = generatePrologFacts(Poll);
- * const result = await perspective.infer(\`
- *   \${facts}
- *   recent_popular_poll(X) :-
- *     poll(X),
- *     poll_vote_count(X, N), N > 10.
- * \`);
- * ```
- *
- * @param ModelClass - A class decorated with `@Model` that extends `Ad4mModel`
- * @returns A multi-line Prolog string ready for use with `perspective.infer()`
- */
-export function generatePrologFacts(ModelClass: typeof Ad4mModel): string {
-  const metadata = ModelClass.getModelMetadata();
-  const predicateName = toSnakeCase(metadata.className);
-  const lines: string[] = [];
-
-  lines.push(`% ${metadata.className} — generated Prolog facts`);
-
-  // Instance recognizer
-  const instanceClause = buildInstanceClause(predicateName, metadata);
-  if (instanceClause) {
-    lines.push("");
-    lines.push(`% Instance recognizer`);
-    lines.push(instanceClause);
-  }
-
-  // Property getters
-  const propClauses = Object.values(metadata.properties)
-    .map((p) => buildPropertyClause(predicateName, p))
-    .filter((c): c is string => c !== null);
-  if (propClauses.length > 0) {
-    lines.push("");
-    lines.push(`% Field getters`);
-    lines.push(...propClauses);
-  }
-
-  // Relation getters
-  const relClauses = Object.values(metadata.relations)
-    .map((c) => buildRelationClause(predicateName, c))
-    .filter((c): c is string => c !== null);
-  if (relClauses.length > 0) {
-    lines.push("");
-    lines.push(`% Relation getters`);
-    lines.push(...relClauses);
-  }
-
-  return lines.join("\n");
-}
+// Re-export generatePrologFacts so existing consumers of './Ad4mModel' are unaffected
+export { generatePrologFacts } from "./prolog-facts";
